Move to layout component folder
